Working with Navigation Guards Guard is used when we don't know whether a user is authenticated or not. Allow users to some routes only if they are authenticated
1st method,
In blog component,
    <Switch>
    {this.state.auth ? <Route path="/new-post" component={NewPost} /> : null}
    <Route path="/posts" component={Posts} />
    <Redirect from="/" to="/posts" />
    </Switch>
If user is not authenticated, the '/new-post' route is not rendered. So, on clicking 'new-post' in navigation, the 'Redirect' route catches all other routes, which are not catched by any of above route & redirects it to '/posts'. This is a guard in React.

2nd method,
In componentDidMount() in NewPost,
    componentDidMount () {
      // if unauth => this.props.history.replace('/posts');
      console.log(this.props);
    }
We check whether a user is authenticated & redirected to '/posts' if not authenticated as shown above.